Add Home navigation on Home tabPress
Previously, if Home tab was pressed while Toolbox Editor tab was focused, and a ToolsStack Screen had been visible, that same ToolsStack Screen would be visible again.

The goal is for ZoneButtons to be visible any time the Home tab is pressed while Toolbox Editor is focused.

Currently, the ToolsStack Screen is still visible for a short time before navigating back to ZoneButtons